+++ orca_step/tk_energy.py
@@ -154,6 +154,8 @@ class TkEnergy(seamm.TkNode):
 
         for key in (
             "auxiliary basis",
+            "grid",
+            "scf convergence",
             "extra keywords",
             "bond orders",
             "Hirshfeld charges",